Renaming some json files and moving tests
